refactor: Simplify BlockProperties.vue by moving each section code to separate files
